refactor: core chat-screens refactor - pages, components, mocks, and utils
