+++ web/apps/nextjs__raw_krabby_chat_api_integrations/app/(routes)/(chat)/mocks/index.tsx
@@ -1,0 +1,108 @@
+export const mockChats = [
+  {
+    id: '1',
+    name: 'Amara Diallo',
+    lastMessage: 'Are you free to hop on a call later?',
+    time: '10:42 AM',
+    unread: 3,
+    online: true,
+  },
+  {
+    id: '2',
+    name: 'Kofi Mensah',
+    lastMessage: 'The build is passing now, finally.',
+    time: '9:17 AM',
+    unread: 0,
+    online: false,
+  },
+  {
+    id: '3',
+    name: 'Fatima Al-Rashid',
+    lastMessage: 'Send me the updated docs when ready.',
+    time: 'Yesterday',
+    unread: 1,
+    online: true,
+  },
+  {
+    id: '4',
+    name: 'Luca Ferretti',
+    lastMessage: 'Looks good to me. Ship it.',
+    time: 'Yesterday',
+    unread: 0,
+    online: false,
+  },
+  {
+    id: '5',
+    name: 'Yuna Park',
+    lastMessage: 'I left a comment on the PR.',
+    time: 'Mon',
+    unread: 0,
+    online: true,
+  },
+  {
+    id: '6',
+    name: 'Dev — Backend',
+    lastMessage: 'Kofi: migrations ran clean ✓',
+    time: 'Mon',
+    unread: 7,
+    online: false,
+  },
+  {
+    id: '7',
+    name: 'Omar Sy',
+    lastMessage: 'Will ping you in the morning.',
+    time: 'Sun',
+    unread: 0,
+    online: false,
+  },
+  {
+    id: '8',
+    name: 'Ingrid Holm',
+    lastMessage: 'Done. Closing the ticket.',
+    time: 'Sun',
+    unread: 0,
+    online: true,
+  },
+];
+
+export const mockMessages = [
+  { id: 'm1', from: 'them', text: 'Hey! Are you free to hop on a call later?', time: '10:38 AM' },
+  { id: 'm2', from: 'me', text: 'Hey Amara! Yeah, what time works for you?', time: '10:39 AM' },
+  {
+    id: 'm3',
+    from: 'them',
+    text: 'Maybe around 3pm? I want to walk through the new API design with you.',
+    time: '10:39 AM',
+  },
+  {
+    id: 'm4',
+    from: 'me',
+    text: "3pm works perfectly. I'll send a calendar invite.",
+    time: '10:40 AM',
+  },
+  {
+    id: 'm5',
+    from: 'them',
+    text: 'Perfect. Also — did you get a chance to look at the auth flow I pushed yesterday?',
+    time: '10:40 AM',
+  },
+  {
+    id: 'm6',
+    from: 'me',
+    text: 'Yes! Looks really clean. I left a couple of minor comments but overall it is solid work.',
+    time: '10:41 AM',
+  },
+  {
+    id: 'm7',
+    from: 'them',
+    text: 'Thanks. I addressed most of them already. The token refresh logic was a bit tricky.',
+    time: '10:41 AM',
+  },
+  {
+    id: 'm8',
+    from: 'me',
+    text: 'Agreed. The sliding window approach you used is clever though.',
+    time: '10:42 AM',
+  },
+  { id: 'm9', from: 'them', text: 'Are you free to hop on a call later?', time: '10:42 AM' },
+];
