+++ web/apps/nextjs__raw_krabby_chat_api_integrations/app/(routes)/(chat)/page.tsx
@@ -1,210 +1,18 @@
 import React from 'react';
 import Link from 'next/link';
-
-const mockChats = [
-  {
-    id: '1',
-    name: 'Amara Diallo',
-    lastMessage: 'Are you free to hop on a call later?',
-    time: '10:42 AM',
-    unread: 3,
-    online: true,
-  },
-  {
-    id: '2',
-    name: 'Kofi Mensah',
-    lastMessage: 'The build is passing now, finally.',
-    time: '9:17 AM',
-    unread: 0,
-    online: false,
-  },
-  {
-    id: '3',
-    name: 'Fatima Al-Rashid',
-    lastMessage: 'Send me the updated docs when ready.',
-    time: 'Yesterday',
-    unread: 1,
-    online: true,
-  },
-  {
-    id: '4',
-    name: 'Luca Ferretti',
-    lastMessage: 'Looks good to me. Ship it.',
-    time: 'Yesterday',
-    unread: 0,
-    online: false,
-  },
-  {
-    id: '5',
-    name: 'Yuna Park',
-    lastMessage: 'I left a comment on the PR.',
-    time: 'Mon',
-    unread: 0,
-    online: true,
-  },
-  {
-    id: '6',
-    name: 'Dev — Backend',
-    lastMessage: 'Kofi: migrations ran clean ✓',
-    time: 'Mon',
-    unread: 7,
-    online: false,
-  },
-  {
-    id: '7',
-    name: 'Omar Sy',
-    lastMessage: 'Will ping you in the morning.',
-    time: 'Sun',
-    unread: 0,
-    online: false,
-  },
-  {
-    id: '8',
-    name: 'Ingrid Holm',
-    lastMessage: 'Done. Closing the ticket.',
-    time: 'Sun',
-    unread: 0,
-    online: true,
-  },
-];
-
-function getInitials(name: string) {
-  return name
-    .split(' ')
-    .map((n) => n[0])
-    .join('')
-    .toUpperCase()
-    .slice(0, 2);
-}
-
-const ChatsIcon = () => (
-  <svg
-    width="18"
-    height="18"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="1.6"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z" />
-  </svg>
-);
-
-const ContactsIcon = () => (
-  <svg
-    width="18"
-    height="18"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="1.6"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <path d="M17 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2" />
-    <circle cx="9" cy="7" r="4" />
-    <path d="M23 21v-2a4 4 0 0 0-3-3.87" />
-    <path d="M16 3.13a4 4 0 0 1 0 7.75" />
-  </svg>
-);
-
-const CallsIcon = () => (
-  <svg
-    width="18"
-    height="18"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="1.6"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <path d="M22 16.92v3a2 2 0 0 1-2.18 2 19.79 19.79 0 0 1-8.63-3.07A19.5 19.5 0 0 1 4.69 12 19.79 19.79 0 0 1 1.61 3.4 2 2 0 0 1 3.6 1.22h3a2 2 0 0 1 2 1.72c.127.96.361 1.903.7 2.81a2 2 0 0 1-.45 2.11L7.91 8.82a16 16 0 0 0 6.29 6.29l.96-.96a2 2 0 0 1 2.11-.45c.907.339 1.85.573 2.81.7A2 2 0 0 1 22 16.92z" />
-  </svg>
-);
-
-const ArchiveIcon = () => (
-  <svg
-    width="18"
-    height="18"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="1.6"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <polyline points="21 8 21 21 3 21 3 8" />
-    <rect x="1" y="3" width="22" height="5" />
-    <line x1="10" y1="12" x2="14" y2="12" />
-  </svg>
-);
-
-const SettingsIcon = () => (
-  <svg
-    width="18"
-    height="18"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="1.6"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <circle cx="12" cy="12" r="3" />
-    <path d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1-2.83 2.83l-.06-.06a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-4 0v-.09A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83-2.83l.06-.06A1.65 1.65 0 0 0 4.68 15a1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1 0-4h.09A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 2.83-2.83l.06.06A1.65 1.65 0 0 0 9 4.68a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 4 0v.09a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 2.83l-.06.06A1.65 1.65 0 0 0 19.4 9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 0 4h-.09a1.65 1.65 0 0 0-1.51 1z" />
-  </svg>
-);
-
-const ProfileIcon = () => (
-  <svg
-    width="18"
-    height="18"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="1.6"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <path d="M20 21v-2a4 4 0 0 0-4-4H8a4 4 0 0 0-4 4v2" />
-    <circle cx="12" cy="7" r="4" />
-  </svg>
-);
-
-const PlusIcon = ({ size = 15 }: { size?: number }) => (
-  <svg
-    width={size}
-    height={size}
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="1.8"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <line x1="12" y1="5" x2="12" y2="19" />
-    <line x1="5" y1="12" x2="19" y2="12" />
-  </svg>
-);
-
-const SearchIcon = ({ size = 13 }: { size?: number }) => (
-  <svg
-    width={size}
-    height={size}
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <circle cx="11" cy="11" r="8" />
-    <line x1="21" y1="21" x2="16.65" y2="16.65" />
-  </svg>
-);
+import {
+  ChatsIcon,
+  ContactsIcon,
+  ArchiveIcon,
+  CallsIcon,
+  SettingsIcon,
+  ProfileIcon,
+  SearchIcon,
+  PlusIcon,
+} from '../../../trash/Icons';
+import MobileBottomNav from './components/MobileBottomNav';
+import { mockChats } from './mocks';
+import { getInitials } from './utils';
 
 export default function ChatHome() {
   return (
@@ -468,52 +276,7 @@ export default function ChatHome() {
       </div>
 
       {/* ── Mobile bottom nav bar ── */}
-      <nav className="md:hidden border-t border-black/15 bg-white shrink-0 flex items-center justify-around px-2 py-2 safe-b">
-        <button
-          title="Chats"
-          className="flex flex-col items-center gap-1 px-4 py-1.5 cursor-pointer"
-        >
-          <span className="text-black">
-            <ChatsIcon />
-          </span>
-          <span className="text-[9px] font-semibold tracking-[0.12em] uppercase text-black">
-            Chats
-          </span>
-        </button>
-        <button
-          title="Contacts"
-          className="flex flex-col items-center gap-1 px-4 py-1.5 cursor-pointer"
-        >
-          <span className="text-black/35">
-            <ContactsIcon />
-          </span>
-          <span className="text-[9px] font-semibold tracking-[0.12em] uppercase text-black/35">
-            Contacts
-          </span>
-        </button>
-        <button
-          title="Calls"
-          className="flex flex-col items-center gap-1 px-4 py-1.5 cursor-pointer"
-        >
-          <span className="text-black/35">
-            <CallsIcon />
-          </span>
-          <span className="text-[9px] font-semibold tracking-[0.12em] uppercase text-black/35">
-            Calls
-          </span>
-        </button>
-        <button
-          title="Settings"
-          className="flex flex-col items-center gap-1 px-4 py-1.5 cursor-pointer"
-        >
-          <span className="text-black/35">
-            <SettingsIcon />
-          </span>
-          <span className="text-[9px] font-semibold tracking-[0.12em] uppercase text-black/35">
-            Settings
-          </span>
-        </button>
-      </nav>
+      <MobileBottomNav />
     </div>
   );
 }
